# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Taawun API
  description: Collaborative workspace management platform API (v2.1 RBAC)
  version: 2.1.0
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register new user (sets token & csrf cookies)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400': { description: Bad Request }
        '409': { description: User already exists }
  /api/v1/auth/login:
    post:
      summary: Login user (sets token & csrf cookies)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': { description: Invalid credentials }
...
  /api/v1/workspaces:
    get:
      summary: List user's workspaces
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
    post:
      summary: Create workspace (Admin/Architect only)
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceRequest'
      responses:
        '201': { description: Workspace created }
        '403': { description: Forbidden (Insufficient role) }
...
components:
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      description: Must match the csrf_token cookie
      schema: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [admin, architect, maintainer, viewer] }
        created_at: { type: string }
    Workspace:
      type: object
      properties:
        id: { type: string }
        owner_id: { type: string }
        name: { type: string }
        description: { type: string }
        current_user_role: { type: string, enum: [architect, maintainer, viewer, admin] }
        created_at: { type: string }
        updated_at: { type: string }
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email: { type: string }
        password: { type: string }
        name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
    WorkspaceRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
...